CG-31 Highlight Tiles Update and File Restructure
