+++ index.js
@@ -1,11 +1,12 @@
 import { createChessboard, initialChessboardRender, reRenderChessboard } from "./src/board/chessboard.js";
 import { initialiseEachPiece } from "./src/pieces/initialise-pieces.js";
-import { removeAllHighlightClasses, highlightTileOnly, pressedTile } from "./src/util/clickedPiece.js";
-import { movePiece, pieceOrTile, tileFullLocation } from "./src/util/movement/movePiece.js";
-import { PLAYERGAME, WHITEPLAYER, BLACKPLAYER } from "./player.js";
-import { clickedEnemyPiece} from "./src/util/management/gameManagement.js";
-import { checkmate } from "./src/util/checkmate/checkmate.js";
-import { gameHasEnded } from "./src/util/management/managementDesign.js";
+import { removeAllHighlightClasses, highlightTileOnly, pressedTile, pieceOrTile } from "./src/util/clickedPiece.js";
+import { movePiece } from "./src/app.js";
+import { PLAYERGAME} from "./src/functions/game_management/player.js";
+import { clickedEnemyPiece} from "./src/functions/game_management/gameManagement.js";
+import { checkmate } from "./src/functions/checkmate/checkmate.js";
+import { gameHasEnded } from "./src/functions/game_management/managementDesign.js";
+import { tileFullLocation } from "./src/util/pieceTileLocation.js";
  
 // Create and render the chessboard
 export const chessBoard = createChessboard();
@@ -45,12 +46,12 @@ const tilePressed = async (tileCheck, event) => {
     // User initially presses on a piece
     } else if (!tileCheck && !initialSelectedPiece) {
         initialSelectedPieceLocation = tileFullLocation(event);
-        initialSelectedPiece = await pressedTile(event, chessBoard);
+        initialSelectedPiece = pressedTile(event, chessBoard);
 
         clickedEnemyPiece(initialSelectedPiece, event);
     // User pressed on a piece and decides to move it
     } else {
-        await movePiece(initialSelectedPiece, initialSelectedPieceLocation, event, chessBoard);
+        movePiece(initialSelectedPiece, initialSelectedPieceLocation, event, chessBoard);
         reRenderChessboard(chessBoard)
 
         checkmate(PLAYERGAME.currentTurn, KINGS, chessBoard);
@@ -64,7 +65,7 @@ chessBoardElement.addEventListener('click', async (event) => {
     if (PLAYERGAME.status !== "finished") {
         const tileCheck = pieceOrTile(event);
 
-        await tilePressed(tileCheck, event);
+        tilePressed(tileCheck, event);
 
         if (PLAYERGAME.status == "finished") {
             gameHasEnded();
