+++ src/functions/checkmate/movingIntoCheck.js
