--- src/util/movement/pathCheck.js
@@ -1,28 +1,0 @@
-const pieceInTileCheck = (inbetweenPositions) => {
-    return inbetweenPositions.pieceInWayCheck;
-}
-
-// Pawn, Rooks, Queen, King
-const horizontalUpCheck = (startingPosition, destinationPosition) => {
-    
-}
-
-// Queen and King
-const horizontalDownCheck = (startingPosition, destinationPosition) => {
-
-}
-
-// Rooks, Queen
-const verticalCheck = (startingPosition, destinationPosition) => {
-
-}
-
-// Bishop, Queen
-const diagonalCheck = (startingPosition, destinationPosition) => {
-
-}
-
-// Pawn Capture
-const pawnCaptureCheck = (startingPosition, destinationPosition) => {
-
-}
