--- src/util/pieceToTile.js
@@ -1,6 +1,0 @@
-export const resetTileData = (tile) => {
-    tile.pieceOnTile = null;
-    tile.tileOwnership = null;
-
-    console.log(`Resetted ${tile.position}`)
-}
